restore DataFrame suite
